requirements.txt
alembic==1.0.0
aniso8601==3.0.2
asn1crypto==0.24.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
click==6.7
clickclick==1.2.2
connexion==1.1.9
cryptography==2.3.1
Flask==1.0.2
Flask-API==1.0
Flask-Migrate==2.2.1
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
idna==2.7
inflection==0.3.1
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
Mako==1.0.7
MarkupSafe==1.0
mysqlclient==1.3.12
pycparser==2.19
PyMySQL==0.9.2
python-dateutil==2.6.0
python-editor==1.0.3
pytz==2018.5
PyYAML==3.13
requests==2.19.1
six==1.11.0
SQLAlchemy==1.2.12
swagger-spec-validator==2.4.0
urllib3==1.23
virtualenv==16.0.0
Werkzeug==0.14.1
